🐛 Manifest and icon not working in production
NextJS is meant to create the relevant link tags
in the head of the document for the manifest and
icon automatically which is working fine in
development but not production.

I suspect this is to do with using an export build
and serving statically rather than using the usual
NextJS magic.