✨ Simpler pagination and 'scroll to' behaviour
The pagination now uses buttons instead of the
infinite scroll which had issues such as the
functionality not working unless there was already
a scrollbar present.

Also updates the 'scroll to' behaviour. There is
no longer an initial scroll on load as the log
limit will always be set low enough that you can
see your next todo. The 's' shortcut is updated
so that it scrolls to the next todo.